Move `ClientDatabase::read_page_groups()` into `StorageBackendAdapter::read_page_groups()`
